apps/object-counter/server/src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.5
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
orjson==3.10.7
packaging==24.1
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.5
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
timm==1.0.9
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
transformers==4.41.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
ultralytics